🚚 rename KnowledgeProvider into KnowledgeToFeedProvider
