+++ frontend/app/brains-management/layout.tsx
@@ -1,7 +1,6 @@
 "use client";
 import { ReactNode } from "react";
 
-import { KnowledgeProvider } from "@/lib/context";
 import { useSupabase } from "@/lib/context/SupabaseProvider";
 import { redirectToLogin } from "@/lib/router/redirectToLogin";
 
@@ -20,7 +19,7 @@ const Layout = ({ children }: LayoutProps): JSX.Element => {
   return (
     <div className="relative h-full w-full flex justify-stretch items-stretch">
       <BrainsList />
-      <KnowledgeProvider>{children}</KnowledgeProvider>
+      {children}
     </div>
   );
 };
