+++ frontend/lib/components/KnowledgeToFeedInput/KnowledgeToFeedInput.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from "react-i18next";
 
 import Button from "@/lib/components/ui/Button";
-import { useKnowledgeContext } from "@/lib/context/KnowledgeProvider/hooks/useKnowledgeContext";
+import { useKnowledgeToFeedContext } from "@/lib/context/KnowledgeToFeedProvider/hooks/useKnowledgeToFeedContext";
 
 import { FeedItems } from "./components";
 import { Crawler } from "./components/Crawler";
@@ -14,7 +14,7 @@ export const KnowledgeToFeedInput = ({
 }): JSX.Element => {
   const { t } = useTranslation(["translation", "upload"]);
 
-  const { knowledgeToFeed } = useKnowledgeContext();
+  const { knowledgeToFeed } = useKnowledgeToFeedContext();
 
   return (
     <div className="px-20">
