--- frontend/lib/context/KnowledgeProvider/index.ts
@@ -1,1 +1,0 @@
-export * from "./knowledge-provider";
