+++ frontend/lib/context/KnowledgeToFeedProvider/hooks/useKnowledgeToFeedContext.tsx
@@ -2,11 +2,11 @@ import { useContext } from "react";
 
 import { FeedItemType } from "@/app/chat/[chatId]/components/ActionsBar/types";
 
-import { KnowledgeContext } from "../knowledge-provider";
+import { KnowledgeToFeedContext } from "../knowledgeToFeed-provider";
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export const useKnowledgeContext = () => {
-  const context = useContext(KnowledgeContext);
+export const useKnowledgeToFeedContext = () => {
+  const context = useContext(KnowledgeToFeedContext);
 
   const addKnowledgeToFeed = (knowledge: FeedItemType) => {
     context?.setKnowledgeToFeed((prevKnowledge) => [
@@ -25,7 +25,7 @@ export const useKnowledgeContext = () => {
   };
 
   if (context === undefined) {
-    throw new Error("useKnowledge must be used inside KnowledgeProvider");
+    throw new Error("useKnowledge must be used inside KnowledgeToFeedProvider");
   }
 
   return { ...context, addKnowledgeToFeed, removeKnowledgeToFeed };
