+++ frontend/lib/context/KnowledgeToFeedProvider/knowledgeToFeed-provider.tsx
@@ -3,37 +3,31 @@
 import { createContext, useState } from "react";
 
 import { FeedItemType } from "@/app/chat/[chatId]/components/ActionsBar/types";
-import { Knowledge } from "@/lib/types/Knowledge";
 
-type KnowledgeContextType = {
-  allKnowledge: Knowledge[];
-  setAllKnowledge: React.Dispatch<React.SetStateAction<Knowledge[]>>;
+type KnowledgeToFeedContextType = {
   knowledgeToFeed: FeedItemType[];
   setKnowledgeToFeed: React.Dispatch<React.SetStateAction<FeedItemType[]>>;
 };
 
-export const KnowledgeContext = createContext<KnowledgeContextType | undefined>(
-  undefined
-);
+export const KnowledgeToFeedContext = createContext<
+  KnowledgeToFeedContextType | undefined
+>(undefined);
 
-export const KnowledgeProvider = ({
+export const KnowledgeToFeedProvider = ({
   children,
 }: {
   children: React.ReactNode;
 }): JSX.Element => {
-  const [allKnowledge, setAllKnowledge] = useState<Knowledge[]>([]);
   const [knowledgeToFeed, setKnowledgeToFeed] = useState<FeedItemType[]>([]);
 
   return (
-    <KnowledgeContext.Provider
+    <KnowledgeToFeedContext.Provider
       value={{
-        allKnowledge,
-        setAllKnowledge,
         knowledgeToFeed,
         setKnowledgeToFeed,
       }}
     >
       {children}
-    </KnowledgeContext.Provider>
+    </KnowledgeToFeedContext.Provider>
   );
 };
