+++ frontend/lib/context/index.tsx
@@ -1,4 +1,4 @@
 export * from "./BrainProvider";
 export * from "./ChatProvider";
 export * from "./FeatureFlagProvider";
-export * from "./KnowledgeProvider";
+export * from "./KnowledgeToFeedProvider";
